+++ modules/carto/src/layers/raster-tile-layer.ts
@@ -2,8 +2,8 @@ import {CompositeLayer, CompositeLayerProps, DefaultProps, Layer, LayersList} fr
 import RasterLayer, {RasterLayerProps} from './raster-layer';
 import QuadbinTileset2D from './quadbin-tileset-2d';
 import SpatialIndexTileLayer from './spatial-index-tile-layer';
-import {TilejsonPropType, TilejsonResult} from '../sources/common';
-import {injectAccessToken} from './utils';
+import type {TilejsonResult} from '../sources/types';
+import {injectAccessToken, TilejsonPropType} from './utils';
 
 export const renderSubLayers = props => {
   const tileIndex = props.tile?.index?.q;
