+++ modules/carto/src/layers/utils.ts
@@ -1,3 +1,5 @@
+import type {TilejsonResult} from '../sources/types';
+
 /**
  * Adds access token to Authorization header in loadOptions
  */
@@ -9,3 +11,15 @@ export function injectAccessToken(loadOptions: any, accessToken: string): void {
     };
   }
 }
+
+export const TilejsonPropType = {
+  type: 'object' as const,
+  value: null as null | TilejsonResult,
+  validate: (value: TilejsonResult, propType) =>
+    (propType.optional && value === null) ||
+    (typeof value === 'object' &&
+      Array.isArray(value.tiles) &&
+      value.tiles.every(url => typeof url === 'string')),
+  compare: 2,
+  async: true
+};
